feat: Move to jiff datetime library.
Chrono is now soft-deprecated. Much of the Rust ecosystem is
switching to jiff. These changes move to Jiff for all but the
interface with polars. Polars is still using chrono, and we
cannot drop the chrono dependency until polars does.

There are some minor integer type changes. The biggest change
is the removal of the `Date` enum in favour of just using
`jiff::civil::DateTime` directly. However, this changes the
roundtrip deserlisation/serialisation behaviour losing the
information on whether the user provided a simple date or a
date-time.

We now Use jiff::SignedDuration internally for the
`PywrDuration` this is consistent with the use of 24-hour
days. The `TimeDomain` is updated to remove the assumption
that it has a fixed duration. Instead a helper method
can return the fixed duration *if* there is one.

The conversion tests have been updated to reflect this change.

+++ pywr-core/src/recorders/hdf.rs
@@ -7,7 +7,6 @@ use crate::models::ModelDomain;
 use crate::network::{MetricSetIndex, Network, ResolutionMaps};
 use crate::scenario::{ScenarioDomain, ScenarioIndex};
 use crate::state::State;
-use chrono::{Datelike, Timelike};
 use hdf5_metno::types::StringError;
 use hdf5_metno::{Extents, Group};
 use ndarray::{Array1, s};
@@ -65,7 +64,7 @@ struct Internal {
 #[repr(C)]
 pub struct DateTime {
     index: usize,
-    year: i32,
+    year: i16,
     month: u8,
     day: u8,
     hour: u8,
